requirements.txt
Appium-Python-Client>=1.0.2
beautifulsoup4>=4.9.1
certifi>=2020.6.20
chardet>=3.0.4
coverage>=5.2.1
idna>=2.10
Jinja2>=2.11.2
lxml>=4.5.2
MarkupSafe>=1.1.1
nose>=1.3.7
nose-htmloutput>=0.6.0
nose2>=0.9.2
python-http-client>=3.3.1
requests>=2.24.0
selenium>=3.141.0
sendgrid>=6.4.6
six>=1.15.0
soupsieve>=2.0.1
starkbank-ecdsa>=1.0.0
urllib3>=1.25.10